requirements.txt
openai
requests
beautifulsoup4
pyyaml
ddgs
defusedxml